use protobuf instead of JSON for NATs messages
