+++ krt-template/README.md
@@ -3,21 +3,28 @@
 ## Generate KRT
 
 ```
-$ ./generate_krts.sh
+./scripts/build_krt.sh <VERSION_NAME>
 ```
 
 This will generate the KRT file inside the `build` folder.
 
-## Entrypoint development
-
-TODO: Find out a better way to develop this with autogenerated code.
-
 ## Call to entrypoint
 
-Call to `run_test.sh <runtime-name> <version-name>` to call a gRPC endpoint. 
+Call to `run_test.sh <RUNTIME_NAME> <VERSION_NAME> <WORKFLOW> <NUM_MSGS> <GREETING_NAME>` to call a gRPC endpoint.
 
 Example: Given you have a Runtime named "test" and you upload a krt for version "greeter-v1", run this command:
 
 ```sh
-./run_test.sh test greeter-v1
+./scripts/run_test.sh my-runtime greeter-v1 PyGreet
+./scripts/run_test.sh my-runtime greeter-v1 GoGreet
+./scripts/run_test.sh my-runtime greeter-v1 GoGreet 2
+./scripts/run_test.sh my-runtime greeter-v1 GoGreet 2 Gustavo
+```
+
+## Protobuf
+
+If you want to change the `public_input.proto` file, remember to generate the protobuf code:
+
+```
+./scripts/generate_protobuf_code.sh
 ```
