--- krt-template/go-greeter/krt.yml
@@ -1,16 +1,0 @@
-version: go-greeter-v5
-description: Version for testing.
-entrypoint:
-  proto: public_input.proto
-  image: konstellation/kre-runtime-entrypoint:latest
-
-nodes:
-  - name: greeter
-    image: konstellation/kre-go:latest
-    src: bin/greeter
-
-workflows:
-  - name: greeting-workflow
-    entrypoint: Greet
-    sequential:
-      - greeter
