--- krt-template/go-greeter/public_input.proto
@@ -1,15 +1,0 @@
-syntax = "proto3";
-
-package entrypoint;
-
-message Request {
-  string name = 1;
-}
-
-message Response {
-  string greeting = 1;
-}
-
-service Entrypoint {
-  rpc Greet(Request) returns (Response) {};
-};
