+++ krt-template/greeter/docs/usage.md
@@ -5,7 +5,9 @@
 You can use the grpcurl to test the created version inside the KRE:
 
 ```
-cat "{\"name\":\"John\"}" | grpcurl -plaintext -d @ "$ENTRYPOINT_URL" entrypoint.Entrypoint/Greet
+echo '{"name":"John"}' | grpcurl -plaintext -d @ "$ENTRYPOINT_URL" entrypoint.Entrypoint/PyGreet
+
+echo '{"name":"John"}' | grpcurl -plaintext -d @ "$ENTRYPOINT_URL" entrypoint.Entrypoint/GoGreet
 ```
 
 Back to the [README](./README.md)
