+++ krt-template/greeter/src/go-greeter/go.mod
@@ -1,0 +1,10 @@
+module greeter
+
+go 1.14
+
+require (
+	github.com/golang/protobuf v1.4.2
+	github.com/konstellation-io/kre-runners/kre-go v0.0.0-00010101000000-000000000000
+)
+
+replace github.com/konstellation-io/kre-runners/kre-go => ../../../../runners/kre-go
