--- krt-template/test/greeter.json
@@ -1,3 +1,0 @@
-{
-  "name": "gustavo"
-}
