--- krt-template/test/test_nats_subjects.json
@@ -1,3 +1,0 @@
-{
-  "Greet": "greet-nats-subject"
-}
