--- krt-template/test/test_request.json
@@ -1,3 +1,0 @@
-{
-  "name": "John"
-}
